Add install target, fix newlib
Newlib was not built with the same options as prior ones, so was a little
larger than needed.  Fix.

Add a "rel=2.5.0 subrel=1 ./build install" target that will take and rebuild
newlib w/proper arduino prefix, install it, do the same for HAL, copy other
libs, and update the platform*.json file.

Remove Azure pipelines, it's not fast enough to build this.