Remove full path from __FILE__
Patch the libcpp C preprocessor code to return the basename of a file
instead of the (full) path when __FILE__ occurs in source code.

Makes ESP8266 BINs built a bit smaller and identical when the same sources
are used, even if there is a difference in path.  The ELF files will
still differ and will contain the full paths as before to enable GDB to
work.