Remove some unneeded install steps
Newlib doesn't need to be built and installed to another directory (it
won't be used, even if there).

Remove copy of libgcc.a as well, since it's already in the GCC tools
dir.

Leave the stdc/stdc++exc since we manually choose amongst them in the
link stage for C++ exceptions enabled/disabled.