Use x64 binary for OpenOCD on Win32
See https://github.com/earlephilhower/arduino-pico/issues/478

It seems the IDE is always installing a 32 bit toolchain on Windows,
even for Winx64.  This breaks OpenOCD because it neesds to be 64b to
talk to the 64b libusb.sys driver.

Avoid the issue by installing a 64 bit OpenOCD on Winx32.  All other
bits will remain x32 anmd safe, and only debugging will be affected
if you are still at x32.