Add FLOCK patch, fix FILE I/O, document builds
Re-introduce the file lock patch from earlier.  W/O it the size set
aside for the file lock is 4 bytes, but we need a recursive mutex of
8 bytes so we end up with data corruption.

Bring a copy of the patches and log of all git versions in the
packaged tarballs to make it clear what was/was not included in a
run.