Enable Newlib _POSIX_TIMERS, update GMP (#66)
* gmp version 6.2.1

Because gmp 6.1.2 is removed from GNU repo, it failed to download.
Now it is 6.2.1.

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>

* newlib: define _POSIX_TIMERS to 0 for micro-ROS

micro-ROS uses clock_gettime() and added a custom one when there is no
real posix timers support. The latest newlib missed the func proto and
caused build error. This patch fixed this issue. Tested on both pico
and pico2.

In include/timer.h
    #if defined(_POSIX_TIMERS)
    ...
    int clock_settime (clockid_t clock_id, const struct timespec *tp);
    int clock_gettime (clockid_t clock_id, struct timespec *tp);

Signed-off-by: Thomas Chou <thomas@wytron.com.tw>